extensive docs for math functions
