+++ src/commonMain/kotlin/io/github/offlinebrain/khexagon/math/Math.kt
@@ -1,6 +1,9 @@
 package io.github.offlinebrain.khexagon.math
 
+import io.github.offlinebrain.khexagon.coordinates.AxisPoint
+import io.github.offlinebrain.khexagon.coordinates.Coordinates
 import io.github.offlinebrain.khexagon.coordinates.FractionalHexCoordinates
+import io.github.offlinebrain.khexagon.coordinates.FromHexCoordinates
 import io.github.offlinebrain.khexagon.coordinates.HexCoordinates
 import kotlin.math.*
 
@@ -21,6 +24,16 @@ fun distance(fromQ: Int, fromR: Int, toQ: Int, toR: Int): Int {
     return maxOf(diffQ, diffR, diffS)
 }
 
+/**
+ * Calculates the hexagonal distance from this coordinate to another.
+ *
+ * @param other The other coordinate.
+ * @return The distance.
+ *
+ * @see [io.github.offlinebrain.khexagon.math.distance]
+ */
+fun <T : AxisPoint> T.distance(other: AxisPoint): Int = distance(this.q, this.r, other.q, other.r)
+
 
 /**
  * Calls the provided callback on each point within a hexagonal area defined by provided origin and radius.
@@ -37,6 +50,24 @@ fun circle(originQ: Int = 0, originR: Int = 0, radius: Int, callback: (Int, Int)
             callback(originQ + q, originR + r)
         }
     }
+}
+
+
+/**
+ * Generates a list of coordinates forming a circular region at this coordinate with a given radius.
+ *
+ * @param radius The radius of the circular region.
+ * @return A list of coordinates within the circular region.
+ *
+ * @see [io.github.offlinebrain.khexagon.math.circle]
+ */
+fun <T> T.circle(radius: Int): List<T>
+        where T : AxisPoint, T : FromHexCoordinates<T> {
+    val result = mutableListOf<T>()
+    circle(this.q, this.r, radius) { q, r ->
+        result.add(HexCoordinates.cached(q, r).into())
+    }
+    return result
 }
 
 
@@ -91,7 +122,7 @@ fun bresenhamsLine(startQ: Int, startR: Int, endQ: Int, endR: Int, process: (x: 
     if (dq >= dr && dq >= ds) {
         test = (dq + test) shr 1
 
-        for (i in 0 until dq) {
+        (0 until dq).forEach { i ->
             test -= dr
             q += sq
             if (test < 0) {
@@ -103,7 +134,7 @@ fun bresenhamsLine(startQ: Int, startR: Int, endQ: Int, endR: Int, process: (x: 
     } else if (ds >= dr) {
         test = (ds + test) shr 1
 
-        for (i in 0 until ds) {
+        (0 until ds).forEach { i ->
             test -= dr
             s += ss
             if (test < 0) {
@@ -116,7 +147,7 @@ fun bresenhamsLine(startQ: Int, startR: Int, endQ: Int, endR: Int, process: (x: 
     } else {
         test = (dr + test) shr 1
 
-        for (i in 0 until dr) {
+        (0 until dr).forEach { i ->
             test -= dq
             r += sr
             if (test < 0) {
@@ -129,6 +160,62 @@ fun bresenhamsLine(startQ: Int, startR: Int, endQ: Int, endR: Int, process: (x: 
 }
 
 
+/**
+ * Uses Bresenham's line generation algorithm to create a path from this coordinate to another.
+ *
+ * This function generates a symmetric line - meaning swapping the start and end points won't affect the result.
+ *
+ * @param end The destination coordinate.
+ * @return A list containing coordinates of each point on the line.
+ *
+ * @see [io.github.offlinebrain.khexagon.math.bresenhamsLine]
+ */
+fun <T> T.bresenhamsLine(end: AxisPoint): List<T>
+        where T : AxisPoint, T : FromHexCoordinates<T> {
+    val line = mutableListOf<T>()
+    bresenhamsLine(this.q, this.r, end.q, end.r) { q, r ->
+        line.add(HexCoordinates(q, r).into())
+    }
+    return line
+}
+
+/**
+ * Generates a line of points between two given points in a hexagonal grid.
+ * The line is generated by calculating the distance between the start and end points,
+ * and then interpolating the coordinates for each step along the line.
+ * The provided process function is called for each point along the line.
+ *
+ * @param startQ The 'q' coordinate of the start point.
+ * @param startR The 'r' coordinate of the start point.
+ * @param endQ The 'q' coordinate of the end point.
+ * @param endR The 'r' coordinate of the end point.
+ * @param process A function that's called for each point on the line.
+ */
+fun line(startQ: Int, startR: Int, endQ: Int, endR: Int, process: (x: Int, y: Int) -> Unit) {
+    val distance = distance(startQ, startR, endQ, endR)
+    for (i in 0..distance) {
+        val q = startQ + (endQ - startQ) * i * 1.0f / distance
+        val r = startR + (endR - startR) * i * 1.0f / distance
+        process(q.roundToInt(), r.roundToInt())
+    }
+}
+
+
+/**
+ * Generates a line of coordinates from this coordinate to another.
+ *
+ * @param other The destination coordinate.
+ * @return A list of coordinates marking a path between the two points.
+ */
+fun <T> T.line(other: Coordinates<*>): List<T>
+        where T : Coordinates<T>, T : FromHexCoordinates<T> {
+    val result = mutableListOf<T>()
+    line(this.q, this.r, other.q, other.r) { q, r ->
+        result.add(HexCoordinates.cached(q, r).into())
+    }
+    return result
+}
+
 fun flatHexWidth(radius: Int) = radius * 2
 fun flatHexHeight(radius: Int) = (radius * cos(PI / 6) * 2).toInt()
 
@@ -136,6 +223,15 @@ fun pointyHexWidth(radius: Int) = flatHexHeight(radius)
 
 fun pointyHexHeight(radius: Int) = flatHexWidth(radius)
 
+/**
+ * Converts a point in pixel coordinates to hexagonal coordinates.
+ * This is done by first normalizing the point's coordinates relative to the layout's origin and size,
+ * then applying the layout's orientation matrix to the normalized coordinates.
+ *
+ * @param layout The layout of the hexagonal grid.
+ * @param point The point in pixel coordinates.
+ * @return The point in hexagonal coordinates.
+ */
 fun pixelToHex(layout: Layout, point: Point): FractionalHexCoordinates {
     val m = layout.orientation
     val size = layout.size
@@ -146,6 +242,15 @@ fun pixelToHex(layout: Layout, point: Point): FractionalHexCoordinates {
     return FractionalHexCoordinates(q, r, -q - r)
 }
 
+/**
+ * Converts a point in hexagonal coordinates to pixel coordinates.
+ * This is done by applying the layout's orientation matrix to the hexagonal coordinates,
+ * then scaling and translating the result by the layout's size and origin.
+ *
+ * @param layout The layout of the hexagonal grid.
+ * @param hex The point in hexagonal coordinates.
+ * @return The point in pixel coordinates.
+ */
 fun hexToPixel(layout: Layout, hex: HexCoordinates): Point {
     val m = layout.orientation
     val size = layout.size
@@ -155,6 +260,15 @@ fun hexToPixel(layout: Layout, hex: HexCoordinates): Point {
     return Point(x + origin.x, y + origin.y)
 }
 
+/**
+ * Calculates the offset of a corner of a hexagon.
+ * This is done by first calculating the angle of the corner relative to the layout's start angle,
+ * then scaling a unit vector at that angle by the layout's size.
+ *
+ * @param layout The layout of the hexagonal grid.
+ * @param corner The index of the corner (0-5).
+ * @return The offset of the corner.
+ */
 fun hexCornerOffset(layout: Layout, corner: Int): Point {
     val m = layout.orientation
     val size = layout.size
@@ -162,6 +276,13 @@ fun hexCornerOffset(layout: Layout, corner: Int): Point {
     return Point(size.x * cos(angle).toFloat(), size.y * sin(angle).toFloat())
 }
 
+/**
+ * Calculates the corners of a hexagon in pixel coordinates.
+ *
+ * @param layout The layout of the hexagonal grid.
+ * @param hex The hexagon.
+ * @return A list of points representing the corners of the hexagon.
+ */
 fun polygonCorners(layout: Layout, hex: HexCoordinates): List<Point> {
     val corners = mutableListOf<Point>()
     val center = hexToPixel(layout, hex)
@@ -172,6 +293,13 @@ fun polygonCorners(layout: Layout, hex: HexCoordinates): List<Point> {
     return corners
 }
 
+/**
+ * Rounds a point in fractional hexagonal coordinates to the nearest integer coordinates.
+ *
+ * @param q The 'q' coordinate of the point.
+ * @param r The 'r' coordinate of the point.
+ * @return The point in integer coordinates.
+ */
 fun hexRound(q: Float, r: Float): HexCoordinates {
     var qInt = round(q).toInt()
     var rInt = round(r).toInt()
